Move icon directory up and fix imports
