+++ src/components/HomeLayout/AppLibrariesSection.tsx
@@ -1,11 +1,6 @@
 import React, { FC } from 'react';
 import Link from '@docusaurus/Link';
-import {
-  Identity,
-  SmartContracts,
-  Streams,
-  Stronghold,
-} from '@site/src/common/icons';
+import { Identity, SmartContracts, Streams, Stronghold } from '@site/src/icons';
 
 interface LanguagesProps {
   languages?: { [key: string]: string };
